refactor: extract shared health metrics to lib/health/metrics.js
ย้าย computeBMI/bmiCategoryThai/bpCategory/coerceMeasurementNumber ออกจาก
lib/elderlySchoolDashboard.js เป็น single source ที่ lib/health/metrics.js
เพราะ employee-health ใช้ร่วมด้วย — เตรียมแยกโมดูลโรงเรียนผู้สูงอายุ
