+++ lib/elderlySchoolDashboard.js
@@ -1,4 +1,5 @@
 import Papa from "papaparse";
+import { computeBMI, bmiCategoryThai, bpCategory } from "@/lib/health/metrics";
 
 function normalizeHeader(s) {
   return String(s || "")
@@ -53,17 +54,6 @@ function pickByVisit({ baseIndices, visitIndices }, rowArr, visit) {
   return pickFirstNonEmptyFromIndices(rowArr, [...(baseIndices || []), ...(visitIndices || [])]);
 }
 
-/** แปลงค่าจากฟอร์ม/ชีต/DB เป็นตัวเลข (รองรับ string) — ใช้ร่วมกับ computeBMI และจุดอื่นที่ต้องการความสม่ำเสมอ */
-export function coerceMeasurementNumber(v) {
-  if (v === null || v === undefined) return null;
-  const s = String(v).trim();
-  if (!s) return null;
-  // handle "120/80" style by taking first number only when caller wants it explicitly
-  const cleaned = s.replace(/,/g, "");
-  const n = Number(cleaned);
-  return Number.isFinite(n) ? n : null;
-}
-
 function parseThaiOrIsoDateToISODate(value) {
   if (!value) return null;
   const raw = String(value).trim();
@@ -90,37 +80,6 @@ function parseThaiOrIsoDateToISODate(value) {
   }
 
   return null;
-}
-
-export function computeBMI(weightKg, heightCmOrM) {
-  const w = coerceMeasurementNumber(weightKg);
-  const hRaw = coerceMeasurementNumber(heightCmOrM);
-  if (!w || !hRaw) return null;
-  const hM = hRaw > 3 ? hRaw / 100 : hRaw;
-  if (hM <= 0) return null;
-  const bmi = w / (hM * hM);
-  return Number.isFinite(bmi) ? bmi : null;
-}
-
-export function bmiCategoryThai(bmi) {
-  if (bmi === null || bmi === undefined) return "unknown";
-  if (bmi < 18.5) return "underweight";
-  if (bmi < 23) return "normal";
-  if (bmi < 25) return "overweight";
-  if (bmi < 30) return "obese1";
-  return "obese2";
-}
-
-export function bpCategory(sys, dia) {
-  const s = coerceMeasurementNumber(sys);
-  const d = coerceMeasurementNumber(dia);
-  if (!s || !d) return "unknown";
-  // Hypotension (low BP)
-  if (s < 90 || d < 60) return "low";
-  if (s < 120 && d < 80) return "normal";
-  if (s >= 140 || d >= 90) return "high";
-  // includes elevated + stage1
-  return "risk";
 }
 
 export async function fetchAndParseSheetCSV(csvUrl, { timeoutMs = 15000 } = {}) {
