+++ lib/elderlySchoolImport.js
@@ -1,4 +1,4 @@
-import { computeBMI, bmiCategoryThai, bpCategory } from "@/lib/elderlySchoolDashboard";
+import { computeBMI, bmiCategoryThai, bpCategory } from "@/lib/health/metrics";
 
 function normalizeHeader(s) {
   return String(s || "")
