+++ lib/employeeHealthDashboard.js
@@ -1,5 +1,5 @@
 import Papa from "papaparse";
-import { computeBMI, bmiCategoryThai, bpCategory } from "@/lib/elderlySchoolDashboard";
+import { computeBMI, bmiCategoryThai, bpCategory } from "@/lib/health/metrics";
 
 function normalizeHeader(s) {
   return String(s || "")
